Add files via upload
Aerosol Optical Depth (AOD) is a measure of calculating the aerosol loading over our at-
mosphere. Higher the AOD values, higher is the aerosol loading over a column of the
atmosphere. Through this thesis, we have tried to compare the AODs over India, and our
findings are —Indo-Gangatic Plain (IGP) has the highest columnar loading of aerosols, while
northern India has the least. We have also discussed the seasonal variations of AOD. We
have not done any calculations and simulations at this stage, but provided a critique to the
existing literatures.
There are various instruments and methods of calculating the AODs, satellite based mea-
surement being the most easiest to operate but lacks in accuracy, while the ground-based
measuring devices like sun-photometers provide reasonable results. Also, there are Chemical
Transport Models, which keep in account the transport and transformation of aerosols, and
are compared with each other.